conflict 방지 및 최신화
