+++ app/pages/User.tsx
@@ -1,5 +1,13 @@
 import React, { useState } from "react";
 import "./User.css";
+import type { User } from "../api/pathbook/types/User";
+import type { Route } from "./pages/+types/User";
+
+export async function loader({ request }: Route.LoaderArgs) {
+
+
+  return {};
+}
 
 export default function UserPage() {
   const [activeTab, setActiveTab] = useState("posts");
@@ -10,10 +18,14 @@ export default function UserPage() {
   const [profileImage, setProfileImage] = useState(
     "/app/assets/image/samplepic2.jpg"
   );
+
   const [showBgModal, setShowBgModal] = useState(false);
   const [showProfileModal, setShowProfileModal] = useState(false);
-  const [showDeleteModal, setShowDeleteModal] = useState(false);
   const [selectedFileName, setSelectedFileName] = useState("선택된 파일 없음");
+
+  const [tabData, setTabData] = useState<any[]>([]);
+  const [tabLoading, setTabLoading] = useState(false);
+  const [user, setUser] = useState<User | null>(null);
 
   const handleImageUpload = (
     e: React.ChangeEvent<HTMLInputElement>,
@@ -33,6 +45,16 @@ export default function UserPage() {
         setSelectedFileName(file.name);
       };
       fileReader.readAsDataURL(file);
+    }
+  };
+
+  const handleEditSave = async () => {
+    try {
+      // await new UpdateProfileRequest(nickname, bio).send();
+      setIsEditing(false);
+      alert("프로필이 저장되었습니다.");
+    } catch {
+      alert("프로필 저장에 실패했습니다.");
     }
   };
 
@@ -69,32 +91,6 @@ export default function UserPage() {
     </div>
   );
 
-  const renderDeleteModal = () => (
-    <div className="modal-overlay">
-      <div className="modal-content">
-        <h3>정말 탈퇴하시겠어요?</h3>
-        <p>이 작업은 되돌릴 수 없습니다.</p>
-        <div className="modal-button-group">
-          <button
-            className="modal-delete-button"
-            onClick={() => {
-              alert("회원 탈퇴가 완료되었습니다.");
-              setShowDeleteModal(false);
-            }}
-          >
-            탈퇴하기
-          </button>
-          <button
-            className="modal-cancel-button"
-            onClick={() => setShowDeleteModal(false)}
-          >
-            취소
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-
   const renderContent = () => {
     switch (activeTab) {
       case "posts":
@@ -114,7 +110,13 @@ export default function UserPage() {
     <div className="mypage-container">
       <div
         className="mypage-header"
-        onClick={isEditing ? () => setShowBgModal(true) : () => { /* 이미지 확대 */ }}
+        onClick={
+          isEditing
+            ? () => setShowBgModal(true)
+            : () => {
+                /* 이미지 확대 */
+              }
+        }
         style={{ cursor: "pointer" }}
       >
         <img src={bgImage} alt="배경 이미지" />
@@ -122,7 +124,6 @@ export default function UserPage() {
 
       {showBgModal && renderModal("background")}
       {showProfileModal && renderModal("profile")}
-      {showDeleteModal && renderDeleteModal()}
 
       <div className="mypage-profile">
         <div className="profile-left">
@@ -132,7 +133,13 @@ export default function UserPage() {
               backgroundImage: `url(${profileImage})`,
               backgroundSize: "cover",
             }}
-            onClick={isEditing ? () => setShowProfileModal(true) : () => { /* 이미지 확대 */ }}
+            onClick={
+              isEditing
+                ? () => setShowProfileModal(true)
+                : () => {
+                    /* 이미지 확대 */
+                  }
+            }
           />
           <div className="profile-info">
             {isEditing ? (
@@ -162,44 +169,28 @@ export default function UserPage() {
         <div className="profile-buttons">
           <button
             className="edit-button"
-            onClick={() => setIsEditing((prev) => !prev)}
+            onClick={
+              isEditing ? () => handleEditSave() : () => setIsEditing(true)
+            }
           >
             {isEditing ? "저장" : "프로필 편집하기"}
-          </button>
-          <button
-            className="edit-button delete-button"
-            onClick={() => setShowDeleteModal(true)}
-          >
-            회원탈퇴
           </button>
         </div>
       </div>
 
       <div className="mypage-tabmenu">
-        <div
-          className={`tab-item ${activeTab === "posts" ? "active" : ""}`}
-          onClick={() => setActiveTab("posts")}
-        >
-          작성글
-        </div>
-        <div
-          className={`tab-item ${activeTab === "comments" ? "active" : ""}`}
-          onClick={() => setActiveTab("comments")}
-        >
-          작성 댓글
-        </div>
-        <div
-          className={`tab-item ${activeTab === "likes" ? "active" : ""}`}
-          onClick={() => setActiveTab("likes")}
-        >
-          좋아요
-        </div>
-        <div
-          className={`tab-item ${activeTab === "bookmarks" ? "active" : ""}`}
-          onClick={() => setActiveTab("bookmarks")}
-        >
-          북마크
-        </div>
+        {["posts", "comments", "likes", "bookmarks"].map((tab) => (
+          <div
+            key={tab}
+            className={`tab-item cursor-pointer ${activeTab === tab ? "active" : ""}`}
+            onClick={() => setActiveTab(tab)}
+          >
+            {tab === "posts" && "작성글"}
+            {tab === "comments" && "작성 댓글"}
+            {tab === "likes" && "좋아요"}
+            {tab === "bookmarks" && "북마크"}
+          </div>
+        ))}
       </div>
 
       <div className="mypage-searchbar">
@@ -214,6 +205,6 @@ export default function UserPage() {
 
 const EmptyContent = ({ label }: { label: string }) => (
   <div className="mypage-content">
-    <div className="content-area">{label} (비워둠)</div>
+    <div className="content-area">{label}</div>
   </div>
 );
